+++ askwalking-admin/src/main/java/com/cloud/askwalking/admin/dto/ModifyConfigureApiDTO.java
@@ -27,6 +27,9 @@ public class ModifyConfigureApiDTO implements Serializable {
     @ApiModelProperty(value = "应用名称，调用服务注册名")
     private String applicationName;
 
+    @ApiModelProperty(value = "协议", allowableValues = "RPC,FEIGN")
+    private String protocol;
+
     @ApiModelProperty(value = "API类型", allowableValues = "API,ADMIN,SAAS")
     private String apiType;
