+++ askwalking-admin/src/main/resources/bootstrap.properties
@@ -1,6 +1,5 @@
 spring.main.allow-bean-definition-overriding=true
 spring.application.name=askwalking-admin
 management.endpoints.web.exposure.include=*
-spring.cloud.nacos.config.server-addr=${nacos-server:nacos-server:8848}
+spring.cloud.nacos.config.server-addr=127.0.0.1:8848
 system.code=9001
-logging.path =/logs/askwalking/askwalking-admin
